backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary>=2.9.9

# Background Tasks
celery==5.3.4
redis==5.0.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests>=2.31.0

# File Storage
boto3==1.29.7  # For S3 storage (optional)
Pillow>=10.2.0  # For image processing

# DICOM Processing
pydicom==2.4.4
numpy>=1.26.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket
websockets==12.0

# Utilities
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 3D Model Generation (for generate_3d.py script)
numpy-stl==3.1.0
scikit-image==0.22.0

